LayoutTests/http/tests/media/fairplay/fps-hls-key-rotation-expected.txt


EVENT(canplay)
RUN(video.currentTime = 11)
EVENT(seeked)
END OF TEST
